[TEMPLATES] Rewrite subroutines as operations (#1192)
* transformed approx time evol template

* rewrite arbitrary_unitary

* interferometer tests moved

* port permute

* ucssd ported

* fix qaoa tests

* import gates in files

* black

* update autosummary in docs

* fix test
