+++ pennylane/templates/subroutines/qmc.py
@@ -15,10 +15,9 @@
 Contains the QuantumMonteCarlo template and utility functions.
 """
 import numpy as np
-
 import pennylane as qml
 from pennylane.operation import AnyWires, Operation
-from pennylane.wires import Wires
+from pennylane.ops import QubitUnitary
 
 
 def probs_to_unitary(probs):
@@ -343,8 +342,8 @@ class QuantumMonteCarlo(Operation):
                 "The probability distribution must have a length that is a power of two"
             )
 
-        self.target_wires = Wires(target_wires)
-        self.estimation_wires = Wires(estimation_wires)
+        self.target_wires = list(target_wires)
+        self.estimation_wires = list(estimation_wires)
         wires = self.target_wires + self.estimation_wires
 
         if num_target_wires != len(self.target_wires):
@@ -362,8 +361,8 @@ class QuantumMonteCarlo(Operation):
         A, R, Q = self.parameters
 
         with qml.tape.QuantumTape() as tape:
-            qml.QubitUnitary(A, wires=self.target_wires[:-1])
-            qml.QubitUnitary(R, wires=self.target_wires)
+            QubitUnitary(A, wires=self.target_wires[:-1])
+            QubitUnitary(R, wires=self.target_wires)
             qml.templates.QuantumPhaseEstimation(
                 Q, target_wires=self.target_wires, estimation_wires=self.estimation_wires
             )
