+++ pennylane/templates/subroutines/qpe.py
@@ -16,7 +16,7 @@ Contains the QuantumPhaseEstimation template.
 """
 import pennylane as qml
 from pennylane.operation import AnyWires, Operation
-from pennylane.wires import Wires
+from pennylane.ops import Hadamard, ControlledQubitUnitary, QFT
 
 
 class QuantumPhaseEstimation(Operation):
@@ -108,12 +108,12 @@ class QuantumPhaseEstimation(Operation):
     par_domain = "A"
 
     def __init__(self, unitary, target_wires, estimation_wires, do_queue=True):
-        self.target_wires = Wires(target_wires)
-        self.estimation_wires = Wires(estimation_wires)
+        self.target_wires = list(target_wires)
+        self.estimation_wires = list(estimation_wires)
 
         wires = self.target_wires + self.estimation_wires
 
-        if len(Wires.shared_wires([self.target_wires, self.estimation_wires])) != 0:
+        if any(wire in self.target_wires for wire in self.estimation_wires):
             raise qml.QuantumFunctionError(
                 "The target wires and estimation wires must be different"
             )
@@ -130,11 +130,11 @@ class QuantumPhaseEstimation(Operation):
 
         with qml.tape.QuantumTape() as tape:
             for wire in self.estimation_wires:
-                qml.Hadamard(wire)
-                qml.ControlledQubitUnitary(
+                Hadamard(wire)
+                ControlledQubitUnitary(
                     unitary_powers.pop(), control_wires=wire, wires=self.target_wires
                 )
 
-            qml.QFT(wires=self.estimation_wires).inv()
+            QFT(wires=self.estimation_wires).inv()
 
         return tape
